+++ changelog.md
@@ -3,6 +3,7 @@
 **v2.5.6**  
 1.Added shortcut button to update division status  
 2.Fixed `sigfig` showing `undefined` suffix for number larger than 1e15  
+3.Added show / hide password button for login  
 
 **v2.5.5**  
 1.Removed lazy load